Divided the code into three pieces that is ingest_data train and score
