+++ src/awesome_package/train.py
@@ -1,0 +1,20 @@
+import numpy as np
+from sklearn.metrics import mean_absolute_error, mean_squared_error
+
+
+def score_model_rmse(model, X_test, y_test):
+    predictions = model.predict(X_test)
+    mse = mean_squared_error(y_test, predictions)
+    rmse = np.sqrt(mse)
+    return rmse
+
+
+def score_model_mae(model, X_test, y_test):
+    predictions = model.predict(X_test)
+    mae = mean_absolute_error(y_test, predictions)
+    return mae
+
+
+def RF_score(cvres):
+    for mean_score, params in zip(cvres["mean_test_score"], cvres["params"]):
+        print(np.sqrt(-mean_score), params)
